Move CallbackMixin to util
