+++ resumable/core.py
@@ -1,13 +1,13 @@
 import os
 from enum import Enum
 import uuid
-from collections import defaultdict
 import mimetypes
 
 import requests
 
 from resumable.file import LazyLoadChunkableFile
 from resumable.worker import ResumableWorkerPool
+from resumable.util import CallbackMixin
 
 
 MB = 1024 * 1024
@@ -17,26 +17,6 @@ class ResumableSignal(Enum):
     CHUNK_COMPLETED = 0
     FILE_ADDED = 1
     FILE_COMPLETED = 2
-
-
-class CallbackMixin(object):
-
-    def __init__(self, *args, **kwargs):
-        super(CallbackMixin, self).__init__(*args, **kwargs)
-        self.signal_callbacks = defaultdict(list)
-        self.signal_proxy_targets = []
-
-    def register_callback(self, signal, callback):
-        self.signal_callbacks[signal].append(callback)
-
-    def proxy_signals_to(self, target):
-        self.signal_proxy_targets.append(target)
-
-    def send_signal(self, signal):
-        for callback in self.signal_callbacks[signal]:
-            callback()
-        for target in self.signal_proxy_targets:
-            target.send_signal(signal)
 
 
 class Resumable(CallbackMixin):
