+++ resumable/util.py
@@ -1,3 +1,26 @@
+from collections import defaultdict
+
+
+class CallbackMixin(object):
+
+    def __init__(self, *args, **kwargs):
+        super(CallbackMixin, self).__init__(*args, **kwargs)
+        self.signal_callbacks = defaultdict(list)
+        self.signal_proxy_targets = []
+
+    def register_callback(self, signal, callback):
+        self.signal_callbacks[signal].append(callback)
+
+    def proxy_signals_to(self, target):
+        self.signal_proxy_targets.append(target)
+
+    def send_signal(self, signal):
+        for callback in self.signal_callbacks[signal]:
+            callback()
+        for target in self.signal_proxy_targets:
+            target.send_signal(signal)
+
+
 class FixedUrlSession(object):
     """A simple wrapper for requests.Session that fixes the URL."""
